requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
numpy==1.23.4
    # via pandas
pandas==1.5.1
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2022.5
    # via pandas
pyyaml==6.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
